backend/requirements.txt
deepagents>=0.5.2
langchain>=1.2.15
langgraph>=1.1.6
langchain-anthropic>=1.4.0
langchain-openai>=1.1.12
langchain-google-genai>=4.2.1
fastapi>=0.135.3
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.0.0
